Add UIExplorerApplication and MoviesApplication
Make the Examples follow the structor to the generator.
