+++ Examples/Movies/android/app/src/main/AndroidManifest.xml
@@ -12,6 +12,7 @@
       android:targetSdkVersion="22" />
 
     <application
+      android:name=".MoviesApplication"
       android:allowBackup="true"
       android:label="@string/app_name"
       android:icon="@drawable/rotten_tomatoes_icon"
