+++ Examples/Movies/android/app/src/main/java/com/facebook/react/movies/MoviesActivity.java
@@ -13,56 +13,11 @@
  */
 package com.facebook.react.movies;
 
-import android.os.Bundle;
-
 import com.facebook.react.ReactActivity;
-import com.facebook.react.ReactNativeHost;
-import com.facebook.react.ReactPackage;
-import com.facebook.react.shell.MainReactPackage;
-
-import java.util.Arrays;
-import java.util.List;
-
-import javax.annotation.Nullable;
 
 public class MoviesActivity extends ReactActivity {
-  private ReactNativeHost mReactNativeHost;
-
   @Override
   protected String getMainComponentName() {
     return "MoviesApp";
-  }
-
-  @Override
-  protected void onCreate(Bundle savedInstanceState) {
-    mReactNativeHost = new ReactNativeHost(getApplication()) {
-      @Override
-      public String getJSMainModuleName() {
-        return "Examples/Movies/MoviesApp.android";
-      }
-
-      @Override
-      public @Nullable String getBundleAssetName() {
-        return "MoviesApp.android.bundle";
-      }
-
-      @Override
-      protected boolean getUseDeveloperSupport() {
-        return true;
-      }
-
-      @Override
-      protected List<ReactPackage> getPackages() {
-        return Arrays.<ReactPackage>asList(
-          new MainReactPackage()
-        );
-      }
-    };
-    super.onCreate(savedInstanceState);
-  }
-
-  @Override
-  public ReactNativeHost getReactNativeHost() {
-    return mReactNativeHost;
   }
 }
