+++ Examples/UIExplorer/android/app/src/main/AndroidManifest.xml
@@ -16,6 +16,7 @@
     android:targetSdkVersion="23" />
 
   <application
+      android:name=".UIExplorerApplication"
       android:allowBackup="true"
       android:icon="@drawable/launcher_icon"
       android:label="@string/app_name"
