+++ Examples/UIExplorer/android/app/src/main/java/com/facebook/react/uiapp/UIExplorerApplication.java
@@ -1,0 +1,44 @@
+package com.facebook.react.uiapp;
+
+import android.app.Application;
+
+import com.facebook.react.ReactApplication;
+import com.facebook.react.ReactNativeHost;
+import com.facebook.react.ReactPackage;
+import com.facebook.react.shell.MainReactPackage;
+
+import java.util.Arrays;
+import java.util.List;
+
+import javax.annotation.Nullable;
+
+public class UIExplorerApplication extends Application implements ReactApplication {
+  private final ReactNativeHost mReactNativeHost = new ReactNativeHost(this) {
+    @Override
+    public String getJSMainModuleName() {
+      return "Examples/UIExplorer/UIExplorerApp.android";
+    }
+
+    @Override
+    public @Nullable String getBundleAssetName() {
+      return "UIExplorerApp.android.bundle";
+    }
+
+    @Override
+    public boolean getUseDeveloperSupport() {
+      return true;
+    }
+
+    @Override
+    public List<ReactPackage> getPackages() {
+      return Arrays.<ReactPackage>asList(
+        new MainReactPackage()
+      );
+    }
+  };
+
+  @Override
+  public ReactNativeHost getReactNativeHost() {
+    return mReactNativeHost;
+  }
+};
